% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adj.R
\name{get_adj}
\alias{get_adj}
\title{get_adj() Extract adjacency matrices for given variables from tidygraph objects}
\usage{
get_adj(graph, vars, names = NULL)
}
\description{
This function extracts an array of adjacency matrices, of dimension n x n x k,
where k is the number of variables supplied in \code{myvars}.
}
