% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_format.R
\name{get_format}
\alias{get_format}
\title{get_format() Transform Adjancey Matrices into Modelable Data}
\usage{
get_format(data, type = NULL, myfun = NULL)
}
\arguments{
\item{data}{an array of adjacency matrices, produced by \code{\link[=get_adj]{get_adj()}}.}

\item{type}{(Optional) string describing what preset format to convert matrices to. Options: \code{"edgewise"}, or NULL.}

\item{myfun}{(Optional) name of custom-made function. Used when type is not specified.}
}
\description{
This function transforms an array of adjacency matrices into modelable data using a preset or custom function.
}
