% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\title{get_names() Extract node names from graph}
\usage{
get_names(graph, names = NULL)
}
\description{
This function extracts a designated vector of nodes from a tidygraph object,
or generates a set from 1 to n nodes if not provided.
}
