% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stat.R
\name{get_stat}
\alias{get_stat}
\title{get_stat()}
\usage{
get_stat(model)
}
\arguments{
\item{model}{Named model object}
}
\value{
A tibble of coefficients and goodness of fit statistics
}
\description{
This function takes a model object and returns a tidy data.frame
containing model coefficients and goodness of fit statistics, using the broom package.
}
\examples{

# Load dplyr for pipe
library(dplyr)

# Use mtcars dataset
m1 <- mtcars \%>\%
  # predict horsepower using other covariates
  lm(formula = hp ~ cyl + disp)

# Extract statistics from model m1
m1 \%>\% get_stat()
}
