% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vars.R
\name{get_vars}
\alias{get_vars}
\title{get_vars() Extract variable names from formula}
\usage{
get_vars(formula)
}
\description{
This function extracts the outcome and predictor variables as a character string from any formula.
}
