% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netboot.R
\name{netboot}
\alias{netboot}
\title{netboot() Network Node Bootstrap (underlying function)}
\source{
\url{https://github.com/cran/snowboot/blob/master/src/vertboot_matrix_rcpp.cpp}

Snijders, Tom A.B., and Borgatti, Stephen P. (1999). Non-Parametric Standard Errors and Tests for Network Statistics. Connections 22(2), 61-70.
}
\usage{
netboot(matrix, boot)
}
\arguments{
\item{matrix}{Adjacency matrix (usually as outcome variable)}

\item{boot}{vector of node names/numbers. These should be bootstrapped (resampled with replacement) before use.}
}
\value{
A bootstrapped adjacency matrix
}
\description{
This function implements the network (node) bootstrap algorithm from Snijders and Borgatti (1999),
adapting the snowboot package's Rccp code to work for any supplied set of bootstrapped nodes.
}
\examples{

# Load Packages
library(tidyperm)
library(igraph)
library(tidygraph)

# Import pokemon network
data("pokemon")

a <- pokemon \%>\%
 # Get matrix of 151 first-gen nodes
 filter(generation == 1) \%>\%
 # To extract the matrix
 as.igraph() \%>\%
 # Extract directed matrix, with edges weight by damage.
 as_adjacency_matrix(type = "both", attr = "damage", edges = FALSE, names = TRUE, sparse = FALSE)

# Get bootstrapped sample of names
myboot <- colnames(a) \%>\% sample(replace = TRUE)

# Extract bootstrapped matrix
netboot(matrix = a, boot = myboot)
}
