% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokemon.R
\docType{pokemon}
\name{pokemon}
\alias{pokemon}
\title{Pokemon Network (Tidygraph Object)}
\format{
A tidygraph dataset with 801 nodes and 623,979 edges, containing:

\strong{Nodes}: A node dataset with 801 nodes and 14 variables:

\describe{

\item{\strong{id}: Unique Pokedex ID number for each Pokemon.}

\item{\strong{name}: Unique name of each Pokemon.}

\item{\strong{generation}: Generation in which Pokemon was first introduced.}

\item{\strong{capture_rate}: a number from 0 to 255 used to evaluate likelihood of capturing that Pokemon with a Pokeball.
Higher numbers produce higher likelihoods.}

\item{\strong{type1} & \strong{type2}: respective types of each Pokemon.}

\item{\strong{hp}, \strong{attack}, \strong{defense}, \strong{sp_attack}, \strong{sp_defense}, \strong{speed}: raw base stats at level 1 for each Pokemon.}

\item{\strong{legendary}: binary value indicating whether Pokemon is a legendary Pokemon or not (1 = Yes, 0 = No).}

\item{\strong{image}: the image src link for locating each Pokemon's image in our directory.}

\item{\strong{image_mini}: the download link for each of these images originally.}

\item{\strong{page}: the link to each Pokemon's page on Bulbapedia.}
}

\strong{Edges}: An edgelist data frame with 623,979 rows and 7 variables:

\describe{
\item{\strong{from}: recipient Pokemon, being attacked.}

\item{\strong{to}: opponent Pokemon, attacking the recipient.}

\item{\strong{weakness}: relative weakness of \strong{from} Pokemon to \strong{to} Pokemon,
due to their type(s). Reflects ratio of actual damage received to damage given;
eg. \code{weakness = 2.0} means per unit of damage given, 2 times as much damage is received.
Since, Pokemon only attack with one 'type' at a time, the weakness of Pokemon A to Pokemon B
is calculated using the worst possible pairing of these two types.)}

\item{\strong{damage}: points of damage levied from \strong{to} Pokemon on \strong{from} Pokemon,
from a physical attack (\code{power = 40}), already incorporating stats and weakness.}

\item{\strong{turns}: minimum number of turns it takes for \strong{to} Pokemon
to defeat the \strong{from} Pokemon with a physical attack (\code{power = 40}).}

\item{\strong{damage_sp}: points of damage levied from \strong{to} Pokemon on \strong{from} Pokemon,
from a special attack (\code{power = 40}), already incorporating stats and weakness.}

\item{\strong{turns_sp}: minimum number of turns it takes for \strong{to} Pokemon
to defeat the \strong{from} Pokemon with a special attack (\code{power = 40}).}
}
}
\source{
Network dataset produced by Tim Fraser: \url{https://rpubs.com/timothyfraser/pokemon_matrix/}

Original data by Rounak Banik on Kaggle: \url{https://www.kaggle.com/datasets/rounakbanik/pokemon?resource=download}

Images and damage formula by Bulbapedia contributors: \url{https://bulbapedia.bulbagarden.net/wiki/List_of_Pok\%C3\%A9mon_by_National_Pok\%C3\%A9dex_number}
}
\description{
A tidygraph network dataset, containing Pokemon (generations 1-7) as nodes
and damage dealt by Pokemon \emph{i} to Pokemon \emph{j} as edges.
}
\examples{
# Load in the tidygraph object...
data("pokemon")

# And view the object!
pokemon

}
